chore: move trace to extension trait
