+++ crates/provider/src/provider/mod.rs
@@ -8,7 +8,7 @@ mod sendable;
 pub use sendable::SendableTx;
 
 mod r#trait;
-pub use r#trait::{FilterPollerBuilder, Provider, TraceCallList};
+pub use r#trait::{FilterPollerBuilder, Provider};
 
 mod wallet;
 pub use wallet::WalletProvider;
